update: move recovery component into a common action & apply it in Deposits/Withdraws actions
